chatbot/requirements.txt
elasticsearch
load_dotenv
streamlit
python-dotenv
sentence-transformers
langchain_community
langchain_elasticsearch
langchain_aws
langchain_cohere
langchain-openai
NLTK
lark
openai
langchain
jq
elastic-apm
boto3
psutil